app/requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl
validators
google-api-python-client
gunicorn
uvicorn
fastapi